chore(auth): move common middleware to factory functions
